Research collection, more documentation, comments.
Collected URLs and their results for listing posts and listing
JSON.

Added more fields to item types and added comments.

Added more information to the spider class in the comments for the
URLs to use.

TODO: Rename all fields in Post/Comment to be their direct
equivalents as per the Reddit API.