More unit tests, covering storage and indexing.
We want to be able to use regular SQL to query for things, without
losing the native format of the JSON. Therefore, we store the id
by stripping its prefix and then store it again with the prefix as
prefixed_id. This way, all sorts of queries can be performed
without any weird SQL hacks.

Broken unit tests for now to remind me what I was doing.